requirements.txt
pandas
numpy
kaggle
scipy
snakemd
pyyaml

unrar (command-line tool, not installable through pip)